refactor: remove browsers from homepage
