+++ apps/docs/src/.vitepress/components/BrowsersList.stories.ts
@@ -1,0 +1,14 @@
+import type { Meta, StoryObj } from "@storybook/vue3";
+import Component from "./BrowsersList.vue";
+
+const meta: Meta<typeof Component> = {
+  title: "BrowsersList",
+  component: Component,
+};
+
+export default meta;
+type Story = StoryObj<typeof Component>;
+
+export const Default = {
+  args: {},
+} satisfies Story;
